apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
boto3==1.34.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
jinja2==3.1.2
python-dotenv==1.0.0
pillow==10.1.0
qrcode==7.4.2
pyotp==2.9.0
argon2-cffi==23.1.0
docker==6.1.3
pyyaml==6.0.1
jsonschema==4.20.0
cryptography==41.0.7
PyJWT==2.8.0
prometheus-client==0.19.0
structlog==23.2.0
openai==1.6.1
anthropic
pydantic[email]==2.5.0
langchain
langchain-openai
langchain-anthropic
kubernetes>=28.1.0
